pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "MimeEval"
version = "1.0.0"
description = "Evaluation tools for MIME (Mime Identification Multimodal Evaluation)"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Justin Cho", email = "hjcho@usc.edu" },
    { name = "Michael Saxon", email = "saxon@ucsb.edu" },
    { name = "Tejas Srinivasan", email = "tejass55@usc.edu" }
]
dependencies = [
    "torch==2.2.2",
    "torchvision==0.17.2",
    "pandas==2.2.3",
    "numpy==1.26.4",
    "sentence-transformers==3.4.1",
    "scikit-learn==1.6.1",
    "jsonlines==4.0.0",
    "google-generativeai==0.8.5",
    "av==14.4.0",
    "colorama==0.4.6",
    "loguru==0.7.3",
    "opencv-python==4.11.0.86",
    "win32_setctime==1.2.0",
    "transformers==4.48.2",
    "einops",
    "accelerate",
    "timm",
    "qwen-vl-utils",
    "sentencepiece",
    "boto3==1.36.20",
    "datasets==3.6.0",
    "huggingface-hub==0.28.1",
    "tqdm==4.67.1",
    "pillow==11.1.0",
    "matplotlib==3.10.3",
    "scipy==1.15.1",
    "tokenizers==0.21.0",
    "safetensors==0.5.2",
    "protobuf==5.29.5",
    "requests==2.32.3",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "typing-extensions==4.12.2",
    "packaging==24.2",
    "filelock==3.17.0",
    "fsspec==2025.2.0",
    "multiprocess==0.70.16",
    "dill==0.3.8",
    "pyarrow==20.0.0",
    "networkx==3.4.2",
    "sympy==1.13.3",
    "mpmath==1.3.0",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "fonttools==4.58.1",
    "kiwisolver==1.4.8",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "six==1.17.0",
    "threadpoolctl==3.5.0",
    "joblib==1.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "annotated-types==0.7.0",
    "typing-inspection==0.4.1",
    "propcache==0.3.2",
    "future==1.0.0",
    "ffmpeg-python==0.2.0",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "multidict==6.5.0",
    "frozenlist==1.7.0",
    "async-timeout==5.0.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "aiohappyeyeballs==2.6.1",
    "attrs==25.1.0",
    "charset-normalizer==3.4.1",
    "idna==3.10",
    "urllib3==2.3.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "jinja2==3.1.5",
    "markupsafe==3.0.2",
    "cachetools==5.5.2",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-api-core==2.25.1",
    "google-api-python-client==2.172.0",
    "google-ai-generativelanguage==0.6.15",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.0",
    "grpcio-status==1.71.0",
    "httplib2==0.22.0",
    "proto-plus==1.26.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "rsa==4.9.1",
    "jmespath==1.0.1",
    "s3transfer==0.11.2",
    "botocore==1.36.20",
    "uritemplate==4.2.0",
    "openai==1.90.0",
]

[project.scripts]
mimeeval = "MimeEval.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/MimeEval"] 